JIP-187 대출 페이지     - 검색 및 에러 처리 실시
