+++ src/component/rent/RentModal.js
@@ -1,12 +1,9 @@
-import React, { useState } from "react";
+import React from "react";
 import PropTypes from "prop-types";
 import MidModal from "../utils/MidModal";
 import RentModalConfirm from "./RentModalConfirm";
 import RentModalUser from "./RentModalUser";
 import RentModalBook from "./RentModalBook";
-import MiniModal from "../utils/MiniModal";
-import ModalContentsOnlyTitle from "../utils/ModalContentsOnlyTitle";
-import ModalContentsTitleWithMessage from "../utils/ModalContentsTitleWithMessage";
 
 const RentModal = ({
   selectedUser,
@@ -15,44 +12,14 @@ const RentModal = ({
   setSelectedUser,
   setSelectedBooks,
   setMidModalContents,
+  setMiniModalContents,
 }) => {
-  const [miniModalContents, setMiniModalContents] = useState("");
-  const [rentResult, setRentResult] = useState(false);
-
-  const bookTitle = () => {
-    if (selectedBooks.length === 0) return "";
-    if (selectedBooks.length === 1) return selectedBooks[0].info.title;
-    return `${selectedBooks[0].info.title}\n${selectedBooks[1].info.title}`;
-  };
-
-  const closeMiniModal = () => {
-    setMiniModalContents("");
-    setMidModalContents("");
-    window.location.reload();
-  };
-
   const closeMidModal = () => {
     setMidModalContents("");
   };
 
   return (
     <>
-      {miniModalContents && (
-        <MiniModal closeModal={closeMiniModal}>
-          {rentResult ? (
-            <ModalContentsTitleWithMessage
-              closeModal={closeMiniModal}
-              title="대출이 완료되었습니다."
-              message={bookTitle}
-            />
-          ) : (
-            <ModalContentsOnlyTitle
-              closeModal={closeMiniModal}
-              title={miniModalContents}
-            />
-          )}
-        </MiniModal>
-      )}
       <MidModal closeModal={closeMidModal}>
         {midModalContents === "inquire user" && (
           <RentModalUser
@@ -73,7 +40,6 @@ const RentModal = ({
             selectedBooks={selectedBooks}
             closeModal={closeMidModal}
             setMiniModalContents={setMiniModalContents}
-            setRentResult={setRentResult}
           />
         )}
       </MidModal>
@@ -95,6 +61,7 @@ RentModal.propTypes = {
   setSelectedUser: PropTypes.func.isRequired,
   setSelectedBooks: PropTypes.func.isRequired,
   setMidModalContents: PropTypes.func.isRequired,
+  setMiniModalContents: PropTypes.func.isRequired,
 };
 
 export default RentModal;
